renames dialog
